Factor out logging, mutex handling and "ticks" to separate files
getTicks() had to be in a separate file to avoid a circular dependency between utils.h and utils/mutex.h
