+++ src/axe.c
@@ -23,6 +23,8 @@
 #include "dvb.h"
 #include "minisatip.h"
 #include "utils.h"
+#include "utils/ticks.h"
+
 #include <arpa/inet.h>
 #include <ctype.h>
 #include <errno.h>
