+++ src/dvb.c
@@ -23,6 +23,8 @@
 #include "pmt.h"
 #include "minisatip.h"
 #include "utils.h"
+#include "utils/ticks.h"
+
 #include <arpa/inet.h>
 #include <ctype.h>
 #include <errno.h>
