+++ src/satipc.c
@@ -25,6 +25,8 @@
 #include "dvbapi.h"
 #include "minisatip.h"
 #include "utils.h"
+#include "utils/ticks.h"
+
 #include <arpa/inet.h>
 #include <ctype.h>
 #include <errno.h>
