+++ src/utils.c
@@ -643,8 +643,6 @@ void myfree(void *x, char *f, int l) {
     free(x);
 }
 
-pthread_mutex_t log_mutex;
-
 char *get_current_timestamp(void) {
     static char date_str[200];
     time_t date;
@@ -660,99 +658,6 @@ char *get_current_timestamp(void) {
              day[t->tm_wday], month[t->tm_mon], t->tm_mday, t->tm_hour,
              t->tm_min, t->tm_sec, t->tm_year + 1900);
     return date_str;
-}
-
-char *get_current_timestamp_log(void) {
-    static char date_str[200];
-    struct timeval tv;
-    struct tm *t;
-
-    if (gettimeofday(&tv, NULL))
-        return "01/01 00:00:20";
-    t = localtime(&tv.tv_sec);
-    if (!t)
-        return "01/01 00:00:20";
-    snprintf(date_str, sizeof(date_str), "%02d/%02d %02d:%02d:%02d.%03d",
-             t->tm_mday, t->tm_mon + 1, t->tm_hour, t->tm_min, t->tm_sec,
-             (int)(tv.tv_usec / 1000));
-    return date_str;
-}
-
-void _log(const char *file, int line, const char *fmt, ...) {
-    va_list arg;
-    int len = 0, len1 = 0, both = 0;
-    static int idx, times;
-    char stid[50];
-    static char output[2][2000]; // prints just the first 2000 bytes from
-                                 // the message
-
-    /* Check if the message should be logged */
-    opts.last_log = (char *)fmt;
-
-    stid[0] = 0;
-    if (!opts.no_threads) {
-        pthread_mutex_lock(&log_mutex);
-        snprintf(stid, sizeof(stid) - 2, " %s", thread_name);
-        stid[sizeof(stid) - 1] = 0;
-    }
-
-    if (!fmt) {
-        printf("NULL format at %s:%d !!!!!", file, line);
-        if (!opts.no_threads)
-            pthread_mutex_unlock(&log_mutex);
-        return;
-    }
-    idx = 1 - idx;
-    if (idx > 1)
-        idx = 1;
-    else if (idx < 0)
-        idx = 0;
-    if (opts.file_line && !opts.slog)
-        len1 = snprintf(output[idx], sizeof(output[0]),
-                        "[%s%s] %s:%d: ", get_current_timestamp_log(), stid,
-                        file, line);
-    else if (!opts.slog)
-        len1 = snprintf(output[idx], sizeof(output[0]),
-                        "[%s%s]: ", get_current_timestamp_log(), stid);
-    else if (opts.file_line) {
-        len1 = 0;
-        output[idx][0] = '\0';
-    }
-    /* Write the error message */
-    len = len1 =
-        len1 < (int)sizeof(output[0]) ? len1 : (int)sizeof(output[0]) - 1;
-    both = 0;
-    va_start(arg, fmt);
-    len += vsnprintf(output[idx] + len, sizeof(output[0]) - len, fmt, arg);
-    va_end(arg);
-
-    if (strcmp(output[idx] + len1, output[1 - idx] + len1) == 0)
-        times++;
-    else {
-        if (times > 0) {
-            both = 1;
-            snprintf(output[1 - idx], sizeof(output[0]),
-                     "Message repeated %d times", times);
-        }
-        times = 0;
-    }
-
-    if (both) {
-        if (opts.slog)
-            syslog(LOG_NOTICE, "%s", output[1 - idx]);
-        else
-            puts(output[1 - idx]);
-        both = 0;
-    }
-    if (times == 0) {
-        if (opts.slog)
-            syslog(LOG_NOTICE, "%s", output[idx]);
-        else
-            puts(output[idx]);
-    }
-    fflush(stdout);
-    if (!opts.no_threads)
-        pthread_mutex_unlock(&log_mutex);
 }
 
 int endswith(char *src, char *with) {
@@ -1194,167 +1099,6 @@ char *readfile(char *fn, char *ctype, int *len) {
 
 int closefile(char *mem, int len) { return munmap((void *)mem, len); }
 
-#undef DEFAULT_LOG
-#define DEFAULT_LOG LOG_LOCK
-
-int mutex_init(SMutex *mutex) {
-    int rv;
-    pthread_mutexattr_t attr;
-
-    if (opts.no_threads)
-        return 0;
-    if (mutex->enabled)
-        return 1;
-
-    pthread_mutexattr_init(&attr);
-    pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE);
-
-    if ((rv = pthread_mutex_init(&mutex->mtx, &attr))) {
-        LOG("mutex init %p failed with error %d %s", mutex, rv, strerror(rv));
-        return rv;
-    }
-
-    mutex->create_time = getTick();
-    mutex->enabled = 1;
-    mutex->state = 0;
-    LOG("Mutex init OK %p", mutex);
-    return 0;
-}
-
-__thread SMutex *mutexes[100];
-__thread int imtx = 0;
-
-int mutex_lock1(char *FILE, int line, SMutex *mutex) {
-    int rv;
-    int64_t start_lock = 0;
-    if (opts.no_threads)
-        return 0;
-
-    if (!mutex || !mutex->enabled) {
-        LOGM("%s:%d Mutex not enabled %p", FILE, line, mutex);
-        return 1;
-    }
-    if (mutex->enabled && mutex->state && tid != mutex->tid) {
-        LOGM("%s:%d Locking mutex %p already locked at %s:%d tid %lx", FILE,
-             line, mutex, mutex->file, mutex->line, mutex->tid);
-        start_lock = getTick();
-    } else
-        LOGM("%s:%d Locking mutex %p", FILE, line, mutex);
-    rv = pthread_mutex_lock(&mutex->mtx);
-    if (!mutex->enabled && rv == 0) {
-        pthread_mutex_unlock(&mutex->mtx);
-        LOG("Mutex %p destroyed meanwhile", mutex);
-        return 1;
-    }
-    if (rv != 0) {
-        LOG("Mutex Lock %p failed", mutex);
-        return rv;
-    }
-    if (start_lock > 0) {
-        uint64_t ms = getTick() - start_lock;
-        char *prev_file = "none";
-        int prev_line = -1;
-        if (mutex && mutex->enabled && mutex->file) {
-            prev_file = mutex->file;
-            prev_line = mutex->line;
-        }
-        LOGL(ms > 1000 ? 1 : DEFAULT_LOG,
-             "%s:%d Locked %p after %ld ms, previously locked at: %s, line "
-             "%d",
-             FILE, line, mutex, ms, prev_file, prev_line);
-    }
-    mutex->file = FILE;
-    mutex->line = line;
-    mutex->state++;
-    mutex->tid = tid;
-    mutex->lock_time = getTick();
-
-    mutexes[imtx++] = mutex;
-    return 0;
-}
-int mutex_unlock1(char *FILE, int line, SMutex *mutex) {
-    int rv = -1;
-    if (opts.no_threads)
-        return 0;
-
-    if (!mutex || mutex->enabled) {
-        LOGM("%s:%d Unlocking mutex %p", FILE, line, mutex);
-        if (mutex) {
-            mutex->state--;
-            rv = pthread_mutex_unlock(&mutex->mtx);
-        }
-    } else
-        LOG("%s:%d Unlock disabled mutex %p", FILE, line, mutex);
-
-    if (rv != 0 && rv != 1 && rv != -1) {
-        LOGM("mutex_unlock failed at %s:%d: %d %s", FILE, line, rv,
-             strerror(rv));
-    }
-    if (rv == 0 || rv == 1)
-        rv = 0;
-
-    if (rv != -1 && imtx > 0) {
-        if ((imtx >= 1) && mutexes[imtx - 1] == mutex)
-            imtx--;
-        else if ((imtx >= 2) && mutexes[imtx - 2] == mutex) {
-            mutexes[imtx - 2] = mutexes[imtx - 1];
-            imtx--;
-        } else
-            LOG("mutex_leak: Expected %p got %p", mutex, mutexes[imtx - 1]);
-    }
-    return rv;
-}
-
-int mutex_destroy(SMutex *mutex) {
-    int rv;
-    if (opts.no_threads)
-        return 0;
-    if (!mutex || !mutex->enabled) {
-        LOG("destroy disabled mutex %p", mutex);
-
-        return 1;
-    }
-    mutex->enabled = 0;
-
-    if ((imtx >= 1) && mutexes[imtx - 1] == mutex)
-        imtx--;
-    else if ((imtx >= 2) && mutexes[imtx - 2] == mutex) {
-        mutexes[imtx - 2] = mutexes[imtx - 1];
-        imtx--;
-    }
-
-    if ((rv = pthread_mutex_unlock(&mutex->mtx)) != 1 && rv != 0)
-        LOG("%s: pthread_mutex_unlock 1 failed for %p with error %d %s",
-            __FUNCTION__, mutex, rv, strerror(rv));
-
-    // coverity[use : FALSE]
-    if ((rv = pthread_mutex_unlock(&mutex->mtx)) != 1 && rv != 0)
-        LOG("%s: pthread_mutex_unlock 2 failed for %p with error %d %s",
-            __FUNCTION__, mutex, rv, strerror(rv));
-
-    LOG("Destroying mutex %p", mutex);
-    return 0;
-}
-
-void clean_mutexes() {
-    int i;
-    if (!imtx)
-        return;
-    if (opts.no_threads)
-        return;
-    //	LOG("mutex_leak: unlock %d mutexes", imtx);
-    for (i = imtx - 1; i >= 0; i--) {
-        if (!mutexes[i] || !mutexes[i]->enabled)
-            continue;
-        LOG("mutex_leak: %s unlocking mutex %p from %s:%d", __FUNCTION__,
-            mutexes[i], mutexes[i]->file, mutexes[i]->line);
-        mutex_unlock(mutexes[i]);
-    }
-    imtx = 0;
-}
-#undef DEFAULT_LOG
-#define DEFAULT_LOG LOG_UTILS
-
 pthread_t get_tid() { return pthread_self(); }
 
 pthread_t start_new_thread(char *name) {
@@ -1411,27 +1155,6 @@ int add_new_lock(void **arr, int count, int size, SMutex *mutex) {
         }
     mutex_unlock(mutex);
     return -1;
-}
-
-int64_t init_tick;
-
-int64_t getTick() { // ms
-    int64_t theTick;
-    struct timespec ts;
-    clock_gettime(CLOCK_MONOTONIC, &ts);
-    theTick = ts.tv_nsec / 1000000;
-    theTick += ts.tv_sec * 1000;
-    if (init_tick == 0)
-        init_tick = theTick;
-    return theTick - init_tick;
-}
-
-int64_t getTickUs() {
-    int64_t utime;
-    struct timespec ts;
-    clock_gettime(CLOCK_MONOTONIC, &ts);
-    utime = ((int64_t)ts.tv_sec) * 1000000 + ts.tv_nsec / 1000;
-    return utime;
 }
 
 pthread_t join_th[100];
