+++ src/utils/ticks.h
@@ -1,0 +1,10 @@
+#ifndef TICKS_H
+#define TICKS_H
+#define _GNU_SOURCE
+
+#include <stdint.h>
+
+int64_t getTick();
+int64_t getTickUs();
+
+#endif
